chore: 테이블 패키지 변경
