+++ src/test/java/kitchenpos/config/ApplicationTestConfig.java
@@ -5,9 +5,9 @@ import kitchenpos.menu.domain.MenuProductRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.OrderLineItemRepository;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.menu.domain.ProductRepository;
-import kitchenpos.repository.TableGroupRepository;
+import kitchenpos.table.domain.TableGroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
