+++ src/test/java/kitchenpos/table/domain/MenuGroupRepositoryTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.table.domain;
 
 import kitchenpos.config.RepositoryTestConfig;
-import kitchenpos.menu.domain.ProductRepository;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,15 +9,15 @@ import org.springframework.dao.EmptyResultDataAccessException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class ProductRepositoryTest extends RepositoryTestConfig {
+class MenuGroupRepositoryTest extends RepositoryTestConfig {
 
     @Autowired
-    private ProductRepository productRepository;
+    private MenuGroupRepository menuGroupRepository;
 
-    @DisplayName("[EXCEPTION] 상품 식별자값으로 상품을 조회할 수 없는 경우 예외가 발생한다.")
+    @DisplayName("[EXCEPTION] 메뉴 그룹 식별자값으로 메뉴 그룹을 조회하지 못할 경우 예외가 발생한다.")
     @Test
     void throwException_when_notFound() {
-        assertThatThrownBy(() -> productRepository.findProductById(1L))
+        assertThatThrownBy(() -> menuGroupRepository.findMenuGroupById(-1L))
                 .isInstanceOf(EmptyResultDataAccessException.class);
     }
 }
