+++ src/test/java/kitchenpos/table/domain/OrderTableTest.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.table.domain;
 
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
